move data widget to MyAccounts and remove SafeListDrawer
