--- src/components/welcome/SafeListDrawer/styles.module.css
@@ -1,35 +1,0 @@
-.button {
-  padding: var(--space-2);
-}
-
-.buttonText {
-  align-self: center;
-  margin: auto;
-  display: flex;
-  gap: var(--space-1);
-}
-
-.card {
-  margin: auto;
-  padding: var(--space-3);
-  display: flex;
-  flex-direction: column;
-  background-color: transparent;
-}
-
-.card :global .MuiCardHeader-root,
-.card :global .MuiCardContent-root {
-  padding: 0;
-}
-
-.cardHeader {
-  text-align: center;
-}
-
-.infoIcon {
-  vertical-align: middle;
-  width: 1rem;
-  height: 1rem;
-  margin: 4px;
-  color: var(--color-text-secondary);
-}
